+++ examples/cloudflare-worker/src/file.rs
@@ -15,7 +15,7 @@ use http_file::runtime::{AsyncFs, ChunkRead, Meta};
 use xitca_web::bytes::BytesMut;
 
 // folder is relative path pointing at examples/file/static to reuse the static files.
-#[derive(RustEmbed)]
+#[derive(RustEmbed, Clone)]
 #[folder = "../file/static"]
 pub struct Files;