+++ web/src/service/mod.rs
@@ -3,7 +3,7 @@
 #[cfg(feature = "tower-http-compat")]
 pub mod tower_http_compat;
 
-#[cfg(feature = "file")]
+#[cfg(feature = "file-raw")]
 pub mod file;
 
 pub use xitca_service::*;